Refactor: home component file structure
