+++ src/components/homepage/ContentSectionLoading.tsx
